[BugFix] fetchBundle supports double value as timeout.
ResponseHandler now using seconds as params for `wait`, it's needed to support double value as params
so that `fetchBundle` may support waits with Decimals precisely.

issue: f-98263565
AutoSubmit: true

+++ core/runtime/bindings/common/resource/response_handler_proxy.h
@@ -40,7 +40,7 @@ class ResponseHandlerProxy
    * if promise get a std::nullopt, it will return with code
    * `LYNX_BUNDLE_RESOURCE_INFO_TIMEOUT`
    */
-  tasm::BundleResourceInfo WaitAndGetResource(long timeout) {
+  tasm::BundleResourceInfo WaitAndGetResource(double timeout) {
     auto result = promise_->Wait(timeout);
     if (result.has_value()) {
       return *result;
